feat(pptx): render OMML equations; share OMML parser via ooxml-common
PowerPoint stores equations as OMML (ECMA-376 §22.1) — the same grammar
WordprocessingML embeds — wrapped in `a14:m` / `mc:AlternateContent`. The
docx parser already had a complete OMML→MathNode parser and the core math
engine (MathNode→MathML→MathJax SVG) is format-agnostic, so this wires the
same pipeline into pptx and hoists the shared parser so all formats use one
implementation.

- ooxml-common: move docx's `math.rs` here (it only ever matched on element
  local names, so it was already format-agnostic). docx re-exports it. Add
  `Deserialize` (with field defaults) so the AST slots into the pptx type tree.
- pptx parser: extract `m:oMath` / `m:oMathPara` into a new `TextRun::Math`
  variant, descending the `mc:AlternateContent` → `mc:Choice` → `a14:m`
  wrappers and ignoring the rasterized `mc:Fallback`. Markdown export emits
  the flattened equation text.
- core types: add `EquationRun` to the shared `TextRun` union.
- pptx renderer: pre-rasterize each slide's equations to SVG via MathJax
  (`prepareSlideMath`, awaited in `renderSlide`), cache by MathNode[] ref, and
  draw them inline — reserving width, contributing ascent/descent to the line
  box, and centring display math. Equations follow the run/paragraph text
  colour via per-colour `source-in` tinting of the cached glyph image.

Tests: shared OMML tests (24) move with the module; new pptx test covers the
AlternateContent/a14:m unwrapping. Visual fidelity pass pending an
equation-bearing pptx sample.
