+++ packages/ooxml-common/src/lib.rs
@@ -1,10 +1,12 @@
-//! Shared OOXML parsing helpers used by the docx and pptx Rust parsers.
+//! Shared OOXML parsing helpers used by the docx, pptx and xlsx Rust parsers.
 //!
 //! This crate is intentionally small — only logic that already had two
-//! near-duplicate copies in the wild parsers belongs here. Anything
+//! near-duplicate copies in the wild parsers belongs here, plus format-agnostic
+//! grammars like OMML (`math`) that every host schema embeds verbatim. Anything
 //! schema-specific (DocParagraph, ShapeRun, Slide, etc.) stays in the
 //! consuming crate.
 
 pub mod chart;
 pub mod color;
+pub mod math;
 pub mod zip;
